Divided the three sidebar views into separate components. Created Inbox component to house main sidebar elements.
