+++ client/src/components/NewConversation.tsx
@@ -5,6 +5,7 @@ import useSearch from "../hooks/useSearch";
 import Contact from "./Contact";
 import { useAuth } from "../contexts/AuthContext";
 import { useNavigate } from "react-router-dom";
+import { BiArrowBack } from "react-icons/bi";
 
 const NewConversation = () => {
   const [search, setSearch] = useState("");
@@ -16,44 +17,61 @@ const NewConversation = () => {
   const clearSearch = () => setSearch("");
 
   return (
-    <div className="flex flex-col gap-2 flex-1 min-h-0">
-      <div className="flex h-14 justify-center">
-        <Search search={search} setSearch={setSearch} />
+    <>
+      <div className="px-5 py-2 flex items-center justify-between">
+        <div className="flex items-center gap-3">
+          <button
+            onClick={() => navigate(-1)}
+            name="back"
+            className="hover:bg-neutral-200 h-11 aspect-square flex items-center justify-center rounded-full p-2.5 dark:text-white dark:hover:bg-neutral-800"
+            aria-label="Back"
+          >
+            <BiArrowBack size={"1.5rem"} />
+          </button>
+          <h1 className="text-xl font-bold dark:text-white">New message</h1>
+        </div>
       </div>
-      <div className="px-5">
-        <button
-          type="button"
-          onClick={() => navigate("/new/group")}
-          className="w-full mt-2 mb-1 py-2 rounded-lg bg-neutral-200 hover:bg-neutral-300 dark:bg-neutral-800 dark:hover:bg-neutral-700 dark:text-white"
-        >
-          New Group
-        </button>
+      <div className="absolute top-14 left-0 right-0 bottom-0 p-2 flex flex-col justify-between">
+        <div className="flex flex-col gap-2 flex-1 min-h-0">
+          <div className="flex h-14 justify-center">
+            <Search search={search} setSearch={setSearch} />
+          </div>
+          <div className="px-5">
+            <button
+              type="button"
+              onClick={() => navigate("/new/group")}
+              className="w-full mt-2 mb-1 py-2 rounded-lg bg-neutral-200 hover:bg-neutral-300 dark:bg-neutral-800 dark:hover:bg-neutral-700 dark:text-white"
+            >
+              New Group
+            </button>
+          </div>
+          <div className="flex flex-col gap-2 overflow-y-auto flex-1 min-h-0">
+            {!searchResults ? (
+              <p className="px-5 text-center text-neutral-600 dark:text-neutral-500 mt-10">
+                Search for users to start a conversation.
+              </p>
+            ) : searchResults.users.length > 0 ? (
+              searchResults.users.map((result) => {
+                const isCurrentUser = result.id === currentUser?.id;
+                return (
+                  <Contact
+                    img={result.profile_picture || "default-pfp.jpg"}
+                    id={result.id}
+                    displayName={result.display_name}
+                    username={result.username}
+                    key={result.id}
+                    clearSearch={clearSearch}
+                    isCurrentUser={isCurrentUser}
+                  />
+                );
+              })
+            ) : (
+              <h2 className="text-center">No results found</h2>
+            )}
+          </div>
+        </div>
       </div>
-      <div className="flex flex-col gap-2 overflow-y-auto flex-1 min-h-0">
-        {!searchResults ? (
-          <p className="px-5 text-center text-neutral-600 dark:text-neutral-500 mt-10">
-            Search for users to start a conversation.
-          </p>
-        ) : searchResults.users.length > 0 ? (
-          searchResults.users.map((result) => {
-            const isCurrentUser = result.id === currentUser?.id;
-            return (
-              <Contact
-                img={result.profile_picture || "default-pfp.jpg"}
-                id={result.id}
-                displayName={result.display_name}
-                username={result.username}
-                key={result.id}
-                clearSearch={clearSearch}
-                isCurrentUser={isCurrentUser}
-              />
-            );
-          })
-        ) : (
-          <h2 className="text-center">No results found</h2>
-        )}
-      </div>
-    </div>
+    </>
   );
 };
